final version of adding data with api

+++ internal/entity/manager.go
@@ -12,7 +12,7 @@ type (
 
 	Value struct {
 		PriceIndex  float64
-		VolumeIndex float32
+		VolumeIndex float64
 		Coin        *Coin
 		CoinName    string
 		Price       int64
